->auth:otp section fixed ->bundleId changed ->file naming convention followed
