+++ lib/login_screen.dart
@@ -1,10 +1,7 @@
 import 'package:country_code_picker/country_code_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:phone_login/OTPController.dart';
 
-
-
+import 'otp_page.dart';
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({Key? key}) : super(key: key);
@@ -13,8 +10,7 @@ class LoginScreen extends StatefulWidget {
   State<LoginScreen> createState() => _LoginScreenState();
 }
 
-class _LoginScreenState extends State<LoginScreen>
-{
+class _LoginScreenState extends State<LoginScreen> {
   String dialCodeDigits = "+00";
   TextEditingController _controller = TextEditingController();
 
@@ -25,33 +21,30 @@ class _LoginScreenState extends State<LoginScreen>
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-
-            SizedBox(height: 100,),
-
-            Padding(
-                padding: const EdgeInsets.only(left: 28.0, right: 28.0),
-                child: Image.asset("images/login.jpg"),
+            SizedBox(
+              height: 100,
             ),
-
+            Padding(
+              padding: const EdgeInsets.only(left: 28.0, right: 28.0),
+              child: Image.asset("images/login.jpg"),
+            ),
             Container(
               margin: EdgeInsets.only(top: 10),
               child: Center(
                 child: Text(
                   "Phone (OTP) Authentication",
-                  style: TextStyle(fontWeight: FontWeight.bold,fontSize: 20),
+                  style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
                 ),
               ),
             ),
-
-
-
-            SizedBox(height: 50,),
-
+            SizedBox(
+              height: 50,
+            ),
             SizedBox(
               width: 400,
               height: 60,
               child: CountryCodePicker(
-                onChanged: (country){
+                onChanged: (country) {
                   setState(() {
                     dialCodeDigits = country.dialCode!;
                   });
@@ -62,44 +55,40 @@ class _LoginScreenState extends State<LoginScreen>
                 favorite: ["+977", "NP", "+1", "US"],
               ),
             ),
-
             Container(
               margin: EdgeInsets.only(top: 10, right: 10, left: 10),
               child: TextField(
                 decoration: InputDecoration(
-                  hintText: "Phone Number",
-                  prefix: Padding(
-                    padding: EdgeInsets.all(4),
-                    child: Text(dialCodeDigits),
-                  )
-                ),
+                    hintText: "Phone Number",
+                    prefix: Padding(
+                      padding: EdgeInsets.all(4),
+                      child: Text(dialCodeDigits),
+                    )),
                 maxLength: 12,
                 keyboardType: TextInputType.number,
                 controller: _controller,
               ),
             ),
-
             Container(
-            margin: EdgeInsets.all(15),
-            width: double.infinity,
-            child: ElevatedButton(
-              onPressed: (){
-                Navigator.of(context).push(MaterialPageRoute(builder: (c) => OTPControllerScreen(
-                  phone: _controller.text,
-                  codeDigits: dialCodeDigits,
-                )));
-
-              },
-              child: Text('Next',style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),),
+              margin: EdgeInsets.all(15),
+              width: double.infinity,
+              child: ElevatedButton(
+                onPressed: () {
+                  Navigator.of(context).push(MaterialPageRoute(
+                      builder: (c) => OtpPage(
+                            phone: _controller.text,
+                            codeDigits: dialCodeDigits,
+                          )));
+                },
+                child: Text(
+                  'Next',
+                  style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+                ),
+              ),
             ),
-            ),
-
-
-
-
           ],
         ),
-      )
-      ,);
+      ),
+    );
   }
 }
